# contracts/openapi.json
{
  "components": {
    "schemas": {
      "ExposureIn": {
        "description": "One exposure registration. Arm \"A\" is the treated recommendation; arm\n\"B\" is the control/neutral exposure (no WinnerRow required). A\nwinner-linked item derives identity from the winner \u2014 caller identity\nfields are ignored there.",
        "properties": {
          "arm": {
            "anyOf": [
              {
                "enum": [
                  "A",
                  "B"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Arm"
          },
          "channel": {
            "default": "",
            "title": "Channel",
            "type": "string"
          },
          "exposure_id": {
            "minLength": 1,
            "title": "Exposure Id",
            "type": "string"
          },
          "item_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Item Id"
          },
          "item_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Item Version"
          },
          "org_id": {
            "default": "",
            "title": "Org Id",
            "type": "string"
          },
          "pro_id": {
            "default": "",
            "title": "Pro Id",
            "type": "string"
          },
          "recommendation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommendation Id"
          },
          "send_status": {
            "default": "unknown",
            "enum": [
              "unknown",
              "pending",
              "confirmed",
              "failed"
            ],
            "title": "Send Status",
            "type": "string"
          },
          "sent_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sent At"
          }
        },
        "required": [
          "exposure_id"
        ],
        "title": "ExposureIn",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HandoffReceipt": {
        "properties": {
          "handoff_id": {
            "title": "Handoff Id",
            "type": "string"
          },
          "idempotency_key": {
            "title": "Idempotency Key",
            "type": "string"
          },
          "status": {
            "enum": [
              "accepted",
              "rejected",
              "duplicate"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "handoff_id",
          "idempotency_key",
          "status"
        ],
        "title": "HandoffReceipt",
        "type": "object"
      },
      "HandoffResponse": {
        "properties": {
          "receipts": {
            "items": {
              "$ref": "#/components/schemas/HandoffReceipt"
            },
            "title": "Receipts",
            "type": "array"
          }
        },
        "required": [
          "receipts"
        ],
        "title": "HandoffResponse",
        "type": "object"
      },
      "LoginRequest": {
        "properties": {
          "password": {
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
      "RunCreate": {
        "properties": {
          "audience_query": {
            "minLength": 1,
            "title": "Audience Query",
            "type": "string"
          },
          "audience_run": {
            "minLength": 1,
            "title": "Audience Run",
            "type": "string"
          },
          "channels": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Channels",
            "type": "array"
          },
          "journey_window": {
            "default": "churn_risk",
            "enum": [
              "churn_risk",
              "churn_risk_open",
              "onboarding",
              "upsell"
            ],
            "title": "Journey Window",
            "type": "string"
          },
          "loop_config": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Loop Config"
          },
          "pro_ids": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Pro Ids",
            "type": "array"
          }
        },
        "required": [
          "pro_ids",
          "audience_query",
          "audience_run",
          "channels"
        ],
        "title": "RunCreate",
        "type": "object"
      },
...
      "RunView": {
        "properties": {
          "audience_query": {
            "title": "Audience Query",
            "type": "string"
          },
          "audience_run": {
            "title": "Audience Run",
            "type": "string"
          },
          "channels": {
            "items": {
              "type": "string"
            },
            "title": "Channels",
            "type": "array"
          },
          "config_version": {
            "title": "Config Version",
            "type": "string"
          },
          "cost_limit_usd": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Cost Limit Usd",
            "type": "string"
          },
          "cost_reserved_usd": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Cost Reserved Usd",
            "type": "string"
          },
          "cost_spent_usd": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Cost Spent Usd",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "journey_window": {
            "title": "Journey Window",
            "type": "string"
          },
          "loop_config": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Loop Config",
            "type": "object"
          },
          "pro_ids": {
            "items": {
              "type": "string"
            },
            "title": "Pro Ids",
            "type": "array"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "stop_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stop Reason"
          }
        },
        "required": [
          "id",
          "status",
          "pro_ids",
          "audience_query",
          "audience_run",
          "channels",
          "config_version",
          "loop_config",
          "cost_limit_usd",
          "cost_reserved_usd",
          "cost_spent_usd",
          "stop_reason",
          "created_at",
          "journey_window"
        ],
        "title": "RunView",
        "type": "object"
      },
      "TouchOutcomeIn": {
        "description": "One observed exposure/outcome record.\n\nLCM Personalization turns Waypoint's theme into an approved Housecall Pro\nSMS and sends it. It is not the measurement authority; outcome sources\nreport exposure and events back to Waypoint.\n\nV3 authority rules: attribution identity (pro, org, item, arm) comes ONLY\nfrom the winner/exposure records \u2014 callers cannot supply it, and any\ncaller-sent identity or returned_* horizon field is dropped at the wire.\nReturn horizons are derived from first_return_at against a confirmed send,\nor resolved by the checkpoint sweep \u2014 never asserted by a caller.",
        "properties": {
          "channel": {
            "default": "",
            "title": "Channel",
            "type": "string"
          },
          "clicked": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clicked"
          },
          "delivered": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delivered"
          },
          "exposure_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exposure Id"
          },
          "first_return_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Return At"
          },
          "org_id": {
            "default": "",
            "title": "Org Id",
            "type": "string"
          },
          "pro_id": {
            "default": "",
            "title": "Pro Id",
            "type": "string"
          },
          "recommendation_id": {
            "minLength": 1,
            "title": "Recommendation Id",
            "type": "string"
          },
          "replied": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Replied"
          },
          "send_confirmed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Send Confirmed At"
          },
          "send_status": {
            "default": "unknown",
            "enum": [
              "unknown",
              "pending",
              "confirmed",
              "failed"
            ],
            "title": "Send Status",
            "type": "string"
          },
          "sent_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sent At"
          },
          "source": {
            "minLength": 1,
            "title": "Source",
            "type": "string"
          },
          "unsubscribed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unsubscribed"
          }
        },
        "required": [
          "recommendation_id",
          "source"
        ],
        "title": "TouchOutcomeIn",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Pathfinder Waypoint V2",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/auth/login": {
      "post": {
        "operationId": "login_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Login Api Auth Login Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login"
      }
    },
    "/api/exposures": {
      "post": {
        "description": "Canonical exposure registration, including neutral/control (arm B)\nexposures with no WinnerRow. Winner-linked identity is derived from\nthe winner; identity is immutable after registration.",
        "operationId": "ingest_exposures_api_exposures_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/ExposureIn"
                },
                "title": "Body",
                "type": "array"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "integer"
                  },
                  "title": "Response Ingest Exposures Api Exposures Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ingest Exposures"
      }
    },
...
    "/api/outcomes": {
      "post": {
        "description": "Observed outcomes keyed by a canonical winner or neutral exposure id.\nLCM Personalization intake acknowledgement is not send confirmation;\nsee waypoint.outcomes for attribution and idempotency rules.",
        "operationId": "ingest_outcomes_api_outcomes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/TouchOutcomeIn"
                },
                "title": "Body",
                "type": "array"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "integer"
                  },
                  "title": "Response Ingest Outcomes Api Outcomes Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ingest Outcomes"
      }
    },
    "/api/runs": {
      "post": {
        "operationId": "create_run_api_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Run"
      }
    },
    "/api/runs/{run_id}": {
      "get": {
        "operationId": "run_detail_api_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Detail"
      }
    },
    "/api/runs/{run_id}/handoff": {
      "post": {
        "operationId": "create_handoff_api_runs__run_id__handoff_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HandoffResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Handoff"
      }
    },
    "/api/runs/{run_id}/kill": {
      "post": {
        "operationId": "kill_run_api_runs__run_id__kill_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Kill Run"
      }
    },
...
  }
}